Move GetModMan button to buttons subfolder
refs TS-1052
